agent/requirements.txt
fastapi==0.114.2
uvicorn[standard]==0.30.6
python-dotenv==1.0.1